BDOG-835: Remove CSRFExceptionsFilter
It has been deprecated for three years, and the original use-cases for
it no longer appear valid. More analysis is available on the JIRA
ticket.

If you still need the functionality offered by this filter it is
recommended that you copy the code into your own codebase (including the
tests) and take ownership of it.