DLK-714 Update to latest library api for play auditing
Introduced version specific behaviour as Play 26 doesn't provide access to the CoordinatedShutdown.

AuditModule is now specific to the version of play26 as there is only (AFAIK) code sharing between the common and the specific version of play.
This does lead to some duplication in the versions for play27 and play28