DLK-762 Fix final audit counter metric
Upgrades play-auditing to include the work around for the final audit counter metric

When more than one metric is published in the same minute (which happens at shutdown)
the metrics are averaged which means the metrics for the audit sequence number will
be incorrect if these are different. This update publishes a separate metric which
only ever has the final number of audits attempted.