DLK-714 Enable writing audit counters to logs
Audit counters are needed to detect audit loss so they need to be more
relible than the audits themselves. For this reason audit counters are
published as audits, logs & metrics.

Publising audit counters as logs means that info level logs must be written
even when the individual microservice has not enabled info level logging.

This change programatially sets the logging level to INFO for the play-auditing
logger.
It does assume that logback is used as the slf4j backend so this would
need to be updated if that changed