DLK-714 Enable writing audit counters to logs
Audit counters are needed to detect audit loss so they need to be more
relible than the audits themselves. For this reason audit counters are
published as audits, logs & metrics.

Publising audit counters as logs means that info level logs must be written
even when the individual microservice has not enabled info level logging.

This change writes directly to stdout, which will then be read by elasticsearch
as usual, but writes using the JsonEncoder so that the format is correct.