dlk-832: (WIP) add datastream-metrics
In this ticket, we're looking to replace the previous telemetry
implementation for auditing with a simpler counter-based version.

This means that AuditCounters have been removed, and replaced by
DatastreamMetrics.  This comprises:

- A trio of counters (success, failure, reject) that increment depending
on datastream's response to a service's request

- a metricsKey that's packaged with the audit data, telling us what
prefix to look for for this service instance.  By extracting this from
our audit data, we can compare the number _expect_ to come through (from
the point of view of the microservice) with the number that actually do.