BDOG-701 Replace MDCPropagatingExecutorServiceConfigurator with MdcPropagatingDispatcherConfigurator
Rather than implicitly propagating the MDC through the executor, which will be lost over async boundaries when the Thread is changed,
this is now copied over explicitly when prepare is called.
This permits preparing the executor before an async boundary, and is what Play often does.
The prepare function is deprecated, but there is no alternative provided yet.